Convert internal usage of channel preference and direction to use strong enum type
Scheduler exposes channel preference as int as Stackless did before it.
Moved away from representing this internally as int

+++ src/PythonCppType.cpp
@@ -21,7 +21,7 @@ void PythonCppType::Decref()
 	Py_DecRef( m_pythonObject );
 }
 
-int PythonCppType::ReferenceCount()
+Py_ssize_t PythonCppType::ReferenceCount()
 {
 	return m_pythonObject->ob_refcnt;
 }
